renamed to CredentialResponseCoordinator
