+++ src/modules/credentials/handlers/OfferCredentialHandler.ts
@@ -1,6 +1,6 @@
 import type { AgentConfig } from '../../../agent/AgentConfig'
 import type { Handler, HandlerInboundMessage } from '../../../agent/Handler'
-import type { CredentialResponseCoordinator } from '../AutoResponse'
+import type { CredentialResponseCoordinator } from '../CredentialResponseCoordinator'
 import type { CredentialRecord } from '../repository/CredentialRecord'
 import type { CredentialService } from '../services'
